LPS-185512 Rename interface SamlHttpRequestUtil to SamlHttpRequestHelper
