+++ modules/dxp/apps/saml/saml-api/bnd.bnd
@@ -3,6 +3,7 @@ Bundle-SymbolicName: com.liferay.saml.api
 Bundle-Version: 7.2.2
 Export-Package:\
 	com.liferay.saml.constants,\
+	com.liferay.saml.helper,\
 	com.liferay.saml.runtime,\
 	com.liferay.saml.runtime.certificate,\
 	com.liferay.saml.runtime.configuration,\
