+++ modules/dxp/apps/saml/saml-api/src/main/java/com/liferay/saml/helper/SamlHttpRequestHelper.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.saml.util;
+package com.liferay.saml.helper;
 
 import com.liferay.saml.runtime.SamlException;
 
@@ -12,7 +12,7 @@ import javax.servlet.http.HttpServletRequest;
 /**
  * @author Michael C. Han
  */
-public interface SamlHttpRequestUtil {
+public interface SamlHttpRequestHelper {
 
 	public String getEntityDescriptorString(
 			HttpServletRequest httpServletRequest)
