+++ modules/dxp/apps/saml/saml-api/src/main/resources/com/liferay/saml/helper/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
