+++ modules/dxp/apps/saml/saml-opensaml-integration/src/main/java/com/liferay/saml/opensaml/integration/internal/helper/SamlHttpRequestHelperImpl.java
@@ -12,14 +12,15 @@
  *
  */
 
-package com.liferay.saml.opensaml.integration.internal.metadata;
+package com.liferay.saml.opensaml.integration.internal.helper;
 
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.util.HttpComponentsUtil;
 import com.liferay.portal.kernel.util.Validator;
+import com.liferay.saml.helper.SamlHttpRequestHelper;
+import com.liferay.saml.opensaml.integration.internal.metadata.MetadataManager;
 import com.liferay.saml.opensaml.integration.internal.util.OpenSamlUtil;
 import com.liferay.saml.runtime.SamlException;
-import com.liferay.saml.util.SamlHttpRequestUtil;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -29,8 +30,8 @@ import org.osgi.service.component.annotations.Reference;
 /**
  * @author Gabriel Santos
  */
-@Component(service = SamlHttpRequestUtil.class)
-public class SamlHttpRequestUtilImpl implements SamlHttpRequestUtil {
+@Component(service = SamlHttpRequestHelper.class)
+public class SamlHttpRequestHelperImpl implements SamlHttpRequestHelper {
 
 	@Override
 	public String getEntityDescriptorString(
