+++ modules/dxp/apps/saml/saml-opensaml-integration/src/test/java/com/liferay/saml/opensaml/integration/internal/metadata/SamlHttpRequestHelperImplTest.java
@@ -8,6 +8,7 @@ package com.liferay.saml.opensaml.integration.internal.metadata;
 import com.liferay.portal.kernel.servlet.HttpMethods;
 import com.liferay.portal.test.rule.LiferayUnitTestRule;
 import com.liferay.saml.opensaml.integration.internal.BaseSamlTestCase;
+import com.liferay.saml.opensaml.integration.internal.helper.SamlHttpRequestHelperImpl;
 
 import org.junit.Assert;
 import org.junit.Before;
@@ -20,7 +21,7 @@ import org.springframework.mock.web.MockHttpServletRequest;
 /**
  * @author Tomas Polesovsky
  */
-public class MetadataManagerImplTest extends BaseSamlTestCase {
+public class SamlHttpRequestHelperImplTest extends BaseSamlTestCase {
 
 	@ClassRule
 	@Rule
@@ -42,7 +43,7 @@ public class MetadataManagerImplTest extends BaseSamlTestCase {
 
 		Assert.assertEquals(
 			"/c/portal/login",
-			_samlHttpRequestUtilImpl.getRequestPath(mockHttpServletRequest));
+			_samlHttpRequestHelperImpl.getRequestPath(mockHttpServletRequest));
 	}
 
 	@Test
@@ -56,7 +57,7 @@ public class MetadataManagerImplTest extends BaseSamlTestCase {
 
 		Assert.assertEquals(
 			"/c/portal/login",
-			_samlHttpRequestUtilImpl.getRequestPath(mockHttpServletRequest));
+			_samlHttpRequestHelperImpl.getRequestPath(mockHttpServletRequest));
 	}
 
 	@Test
@@ -66,10 +67,10 @@ public class MetadataManagerImplTest extends BaseSamlTestCase {
 
 		Assert.assertEquals(
 			"/c/portal/login",
-			_samlHttpRequestUtilImpl.getRequestPath(mockHttpServletRequest));
+			_samlHttpRequestHelperImpl.getRequestPath(mockHttpServletRequest));
 	}
 
-	private final SamlHttpRequestUtilImpl _samlHttpRequestUtilImpl =
-		new SamlHttpRequestUtilImpl();
+	private final SamlHttpRequestHelperImpl _samlHttpRequestHelperImpl =
+		new SamlHttpRequestHelperImpl();
 
 }
